feat(#393: can run python module from scala)
Apparently a Python string list or a Java ArrayList<String> returned
from a Python method implementing a Java interface cannot be converted
correctly to a List<String> in the JVM via py4j.  We work around the
issue by converting the list to a CSV, for now.  Should track down this
issue when we have more time.