fix(#339: SmvStage can be used as group keys)
GraphUtilTest.testCreateGraphJson only fails when sbt is running
multiple tests. It passes if it is the only test that's run.

The most likely cause of this is that sbt's test runner somehow
results in a specific sequence of lazy val initialization that leads
to creation of two different SmvStage instances for the same stage
name.  Exactly how this occurs is still a mystery.

The apparent fix passes the test and prevents future failure when
using SmvStage as a key.