feat(#466: ignore dataset hash tests, that fail in mvn, pass in sbt)
The only explanation for this strangeness is that the mvn plugin for
scala uses a different scala library from the mvn plugin for
scalatest.  That would also explain why some tests in ClassCRCTest
that use the scala compiler to compile code as part of the test would
work in sbt but not in mvn.  Running with `mvn test` results in
NoSuchMethodError.