moduleUrnsForStage should fail loudly when importing a module causes an error
Previously, when a user error (like a typo in an import statement) caused an import in moduleUrnsForStage to fail, the exception would be caught and all the SmvPyModules in that file would be ignored silently. This behavior was confusing. The error is no longer caught, so the application will exit in the event of such an error.