rerun dead-in-the-middle modules (#554)
Previously, if an application was killed while a module was persisting, the module would leave partially written data which would be read as its persisted data the next time an application was run. We now do not write the module's schema until we have written the module's csv successfully, so reading the persisted data will fail in this scenario due to the missing schema. Additionally, if reading the persisted data fails we delete the module's possible output in case there is partially written data.