Fork JVM for tests
Poking around turned out to reveal that the usual solution to
OOM exceptions is to fork tests. Even Spark didn't seem to resolve
the underlying issue which I think is still there, somewhere deep
in Spark's internals (when using Hive).

For now that seems to do the trick i.e. no more lack of Metaspace issues