fix(#1039): byte[] is passed as value between python and java
Consequently, we cannot pass a byte[] pointer from java to python when
reading from a python file object, but need to ask python for the byte
array when it finishes reading, a chunk at a time.

So the read() method needs to return a byte array rather than filling
in one provided by the caller.