Greater flexibility for supplying --jars, and support for merging --driver-class-path spark argument
Before we did not handle the case of when --option=abc with an equals sign. Additionally, when we ran in cluster mode, the absolute path supplied to --driver-class-path did not work since the JAR is placed by Spark at the root of the YARN container, and therefore needs to be referenced by relative path.