generate and validate metadata after DQM validation (#1231)
Previously, metadata was validated as a step of DQM validation. This made it difficult to isolate the time spent generating metadata from the time spent on pure DQM, and the distinction was not very clear. In order to isolate that time spent on pure DQM, we explicitly cached the metadata before DQM validation. However, this meant that any actions on performed on the dataframe (which already had DQM tasks attached) when generating the user metdadata would trigger the accumulators in the DQMState _ahead of validation_, which meant that each DQMTask failure would have been counted 2+ times by the time the time the DQMTask's policy was applied. By validating metadata as a step independent from DQM validation, we can preserve the timing isolation while avoiding the need to cache the metadata before DQM validation - there's no need to generate the metadata before metadata validation. This is also a clearer conceptual division.